Combine SongListItem components
